command: add new build flag '--keep-builddir'
The working directory contains the files of a target OS system to build
and it's the source to create partitions and OS image. With the flag
'--keep-builddir' won't remove the working directory after the target OS
image is built such that one can traverse the file system in the build
machine. By default, the working directory will be removed.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: I00e8488797535410fde6dbd01249ecb8ebc1ae56