Add to create file system with given file list
This patch will create a partition image with given file list in a file.

For example, suppose we have 'vendor.list' that contains files

  $ cat custom/templates/oem/vendor.list
  /etc/hostname

And give the file name 'vendor.list' to 'hooks/partition' with a certain
partition list.

  $ cat custom/templates/oem/hooks/partition
  oem_mkfs_ext4 5 vendor vendor.list
  oem_mkfs_ext4 6 app
  oem_mkfs_ext4 7 lib

The partition labeled by 'vendor' will contains the files as listed in
'vendor.list'.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: Ie4e80762f2d1c43c77dd941213f449dc3792fcbf