fixup: tempory create UUID of root file system
'flash-kernel' checks root file system when updating initramfs and put
its root file system device into 'conf/param.conf'. But when running
ODROID-Stamper, the device for the root file system is not present as a
block device, so it is confused and add '/dev/sda2' which is not
correct, this leads boot failure.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: I83b832ea411f68e0c0491c0e21684331fad7656e